updates based on review comments
