+++ server/helpers/loginHelper.js
@@ -1,0 +1,32 @@
+const User = require("../models/User");
+const generateToken = require("../utils/generateToken");
+const dotenv = require('dotenv').config({path:__dirname+'/./../.env'})
+
+const loginHelper = async (res, email, password) => {
+  const user = await User.findOne({ email });
+
+  if (user && (await user.matchPassword(password))) {
+    const token = generateToken(user._id);
+    const secondsInWeek = 604800;
+
+    res.cookie("token", token, {
+      httpOnly: true,
+      maxAge: secondsInWeek * 1000
+    });
+
+    res.status(200).json({
+      success: {
+        user: {
+          id: user._id,
+          name: user.name,
+          email: user.email
+        }
+      }
+    });
+  } else {
+    res.status(401);
+    throw new Error("Invalid email or password");
+  }
+};
+
+module.exports = loginHelper;
