+++ server/package.json
@@ -7,7 +7,8 @@
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon ./bin/www"
+    "dev": "nodemon ./bin/www",
+    "seeds": "node ./seeds/demoSeed.js"
   },
   "dependencies": {
     "bcryptjs": "^2.4.3",
