+++ server/sample.env
@@ -1,5 +1,5 @@
 JWT_SECRET=
 MONGO_URI=
 TEAM_NAMES=Bonnie,Sina
-DEMO_USER_EMAIL=
-DEMO_USER_PASSWORD=
+DEMO_USER_EMAIL=demo@demo.com
+DEMO_USER_PASSWORD=testertester
